+++ src/include/sof/audio/audio_stream.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause
  *
- * Copyright(c) 2020 Intel Corporation. All rights reserved.
+ * Copyright(c) 2020 - 2023 Intel Corporation. All rights reserved.
  *
  * Author: Karol Trzcinski <karolx.trzcinski@linux.intel.com>
  */
@@ -26,6 +26,7 @@
 #include <rtos/cache.h>
 #include <ipc/stream.h>
 #include <ipc4/base-config.h>
+#include <module/audio/audio_stream.h>
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -33,44 +34,6 @@
 /** \addtogroup audio_stream_api Audio Stream API
  *  @{
  */
-
-/**
- * set of parameters describing audio stream
- * this structure is shared between audio_stream.h and sink/source interface
- * TODO: compressed formats
- */
-struct sof_audio_stream_params {
-	uint32_t id;
-	enum sof_ipc_frame frame_fmt;	/**< Sample data format */
-	enum sof_ipc_frame valid_sample_fmt;
-
-	uint32_t rate;		/**< Number of data frames per second [Hz] */
-	uint16_t channels;	/**< Number of samples in each frame */
-
-	/**
-	 * align_frame_cnt indicates minimum number of frames that satisfies both byte
-	 * align and frame align requirements. E.g: Consider an algorithm that processes
-	 * in blocks of 3 frames configured to process 16-bit stereo using xtensa HiFi3
-	 * SIMD. Therefore with 16-bit stereo we have a frame size of 4 bytes, and
-	 * SIMD intrinsic requirement of 8 bytes(2 frames) for HiFi3 and an algorithim
-	 * requirement of 3 frames. Hence the common processing block size has to align
-	 * with frame(1), intrinsic(2) and algorithm (3) giving us an optimum processing
-	 * block size of 6 frames.
-	 */
-	uint16_t align_frame_cnt;
-
-	/**
-	 * the free/available bytes of sink/source right shift align_shift_idx, the result
-	 * multiplied by align_frame_cnt is the frame count free/available that can meet
-	 * the align requirement.
-	 */
-	uint16_t align_shift_idx;
-
-	bool overrun_permitted; /**< indicates whether overrun is permitted */
-	bool underrun_permitted; /**< indicates whether underrun is permitted */
-
-	uint32_t buffer_fmt; /**< enum sof_ipc_buffer_format */
-};
 
 /**
  * Audio stream is a circular buffer aware of audio format of the data
