+++ src/include/sof/audio/format.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause
  *
- * Copyright(c) 2016 Intel Corporation. All rights reserved.
+ * Copyright(c) 2016 - 2023 Intel Corporation. All rights reserved.
  *
  * Author: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
  *         Liam Girdwood <liam.r.girdwood@linux.intel.com>
@@ -163,41 +163,6 @@ static inline int16_t q_multsr_sat_16x16(int16_t x, int32_t y,
 static inline int32_t sign_extend_s24(int32_t x)
 {
 	return (x << 8) >> 8;
-}
-
-static inline uint32_t get_sample_bytes(enum sof_ipc_frame fmt)
-{
-	switch (fmt) {
-	case SOF_IPC_FRAME_S16_LE:
-		return 2;
-	case SOF_IPC_FRAME_S24_3LE:
-		return 3;
-	case SOF_IPC_FRAME_U8:
-		return 1;
-	default:
-		return 4;
-	}
-}
-
-static inline uint32_t get_sample_bitdepth(enum sof_ipc_frame fmt)
-{
-	switch (fmt) {
-	case SOF_IPC_FRAME_S16_LE:
-		return 16;
-	case SOF_IPC_FRAME_S24_4LE:
-	case SOF_IPC_FRAME_S24_3LE:
-		return 24;
-	case SOF_IPC_FRAME_U8:
-		return 8;
-	default:
-		return 32;
-	}
-}
-
-static inline uint32_t get_frame_bytes(enum sof_ipc_frame fmt,
-				       uint32_t channels)
-{
-	return get_sample_bytes(fmt) * channels;
 }
 
 #endif /* __SOF_AUDIO_FORMAT_H__ */
